Se hacen algunas modificaciones para los mensajes, eliminar de la lista de promociones las que estan expiradas, agregarle una expiracion por defecto a las promociones

+++ BeaconKontackt/app/src/main/java/model/cache/BeaconCache.java
@@ -18,11 +18,11 @@ public class BeaconCache implements Serializable {
     @DatabaseField
     public String message;
     @DatabaseField
-    public Double expiration;
+    public long expiration;
     @DatabaseField
     public String promoId;
     @DatabaseField
-    public double currentDatetime;
+    public long currentDatetime;
     @DatabaseField
     public boolean messageDisplayed;
